Add a statement of business task
This commit when implemented will add an R markdown file for the Bellabeats case study. Sections/details added in this commit are:
- Outline of the data analysis process
- Ask Section

The outline of the data analysis process demonstrates the steps taken for the data analysis of this case study.

The ask section details the business task, stakeholders, and research questions to be answered.